requirements.txt
feedparser
python-dotenv
openai
supabase
python-dateutil
beautifulsoup4
requests
playwright
